cluster module added , logs setup
Remove hardcoded port 7000

Fix mongo version

Remove middleware/logRequest.ts

Fix more docker build issues

- Copy public.pem correctly. Earlier it was getting renamed to
  .build/src/middleware
- Use environment variable PORT instead of hardcoded PORT number in
  server.ts
- Don't expose the port of the dashboard container. This is because we
  want only connection to the outside world to happen through nginx

Scripts added - SSA_PSW

minor changes

remove rm command from nginx dockerfile

Disable nginx for now

Re-enabled nginx

updated app.ts to show correct error description

minor fix

Resolved an issue in logic

changed patch to put

changed patch to put

changed an endpoint

Changed DELETE to POST

minor fixes - handling undefined in imageurl

fixed getNews

+++ src/controllers/body.ts
@@ -162,7 +162,7 @@ export const updateBody = async (
       req.body.imageUrl = req.file.path;
     }
     if (req.body.imageUrl !== null) {
-      if (body.imageUrl.startsWith('media/')) {
+      if (body.imageUrl !== undefined && body.imageUrl.startsWith('media/')) {
         fs.unlinkSync(body.imageUrl);
       }
     }
